refactor: change project layout and move configs to backend/config/
