+++ backend/config/db.yaml
@@ -1,0 +1,5 @@
+DBuser: yourDBuser
+DBpasswd: yourDBpasswd
+DBip: 127.0.0.1
+DBport: 3306
+DBname: yourDBname
