+++ backend/config/session.yaml
@@ -1,0 +1,1 @@
+SessionKey: whatever_you_want
