+++ backend/internal/model/competition.go
